Rename components
- Cange components name
- Change styles name
